requirements.txt
# Core dependencies
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
scipy==1.11.0

# Visualization
plotly==5.17.0
matplotlib==3.7.0
seaborn==0.12.0
geopandas==0.14.0
folium==0.14.0

# Dashboard
streamlit==1.28.0
streamlit-folium==0.15.0
streamlit-aggrid==0.3.4

# Data processing
pyarrow==14.0.0
openpyxl==3.1.2
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.0
tqdm==4.65.0

# Reporting
reportlab==4.0.4
xlsxwriter==3.1.2
fpdf==1.7.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0